fix: wrap quote

+++ core/src/main/kotlin/moe/kurenai/bot/model/bilibili/DynamicInfo.kt
@@ -74,7 +74,7 @@ data class DynamicInfo(
                     @SerialName("pub_time")
                     val pubTime: String,
                     @SerialName("pub_ts")
-                    val pubTs: Int,
+                    val pubTs: Long,
                     val type: String,
                 ) {
